+++ src/utils/index.tsx
@@ -2,6 +2,7 @@ import { useMenuBehavior } from "./useMenuBehavior";
 import { useStickyScroll } from "./useStickyScroll";
 import { useAutoLoop } from "./useAutoLoop";
 import { useAnimatedReview } from "./useAnimatedReview";
+import { useContactForm } from "./useContactForm";
 import useSanityData from "./useSanityQuery";
 
-export { useMenuBehavior, useStickyScroll, useAutoLoop, useAnimatedReview, useSanityData };
+export { useMenuBehavior, useStickyScroll, useAutoLoop, useAnimatedReview, useSanityData, useContactForm };
